Update patch set 1
Patch Set 1: Code-Review+2

Since this is a change on top of the existing decision to have html in python, I am moving this along.
You raise a valid point litong01 while the django precedent of including html in python does not justify continuing this pattern. Feel free to create a bug (and possibly a patch) if you feel strongly about this.

Patch-set: 1
Label: Code-Review=+2