Update patch set 1
Patch Set 1: Code-Review-1

Hi Rob, I'm changing my review of this to a -1. With the changes that Deklan is making to InfluxDB by splitting out the series into multiple time-series, I'm not sure that we'll need this. Let's measure the performance of the new changes and see if this will be required or not.

Patch-set: 1
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1