Update patch set 1
Patch Set 1:

I am trying to understand the motivation for this change beyond better egg packaging. The code in and below the monitoring module is the monasca-ui while the files in enabled/ and local_settings.py are meant as sample configuration files that need to be copied (in the case of enabled) into horizon folders or appended (in the case of local_settings.py) into horizon/openstack_dashboard/local/local_settings.py. 

That is why the config files are consciously segregated from the monitoring code and we use egg to copy them into a different location. Are you proposing a different model for integrating extensions into horizon?

Patch-set: 1
Label: Workflow=0