Update patch set 1
Patch Set 1: Code-Review+1

If a dimension name exists in the first metric but not in the others it will still be shown. Is that the desired result?
As long as the same version of the agent is populating this it won't happen, but over time, there may be cases where the same metric name has different dimensions.

FYI: The pythonic yet hard to read way to do the same thing is
dict(set.intersection(*(set(d.iteritems()) for d in metric[0:]items()))

Patch-set: 1