Update patch set 3
Patch Set 3:

(1 comment)

The problem is that we are paginating only a single page at a time. It wont give you forward results, just the previous ones...I will implement the change when it is a multiple of 10 to not show a even page. I could most likely make a number like 25 for pagination.

Patch-set: 3