Update patch set 6
Patch Set 6:

(2 comments)

Minor comments on it, also when testing this I could only get it to work if I change the DEFAULT_LINKS in views.py nothing happened when changing the local_settings.py. Is that the expected behavior or am I missing something? Sam had the same problem.

Patch-set: 6
Label: Verified=0