FIX the translation for overview page
"OpenStack Services" and "Servers" in overview cannot be translated
properly, because these words are defined in the localsettings
which are the global variables and these variables are changed
directly.
So there is a possibility that this values can be changed when
two or more users access to the Horizon GUI.
This patch creates the deep copy for changing the values in the
localsettings, and can avoid the affect from the other user
operations.

Change-Id: I2cca00a57f5a13902cc868563597ae5d971c0fff
Closes-Bug: #1580389