Update patch set 1
Patch Set 1: Code-Review-1

Shinya,

This is going to be a great change to the UI!  I like the checkboxes, and i'm thinking we should maybe by default have the 'alarm' notification checked when adding a notification?  Also, when i tested this I didn't see any error logs, but the notification assignment never happened?  Is it working for you?

Thanks!
Brad

Patch-set: 1
Label: Code-Review=-1