Update patch set 2
Patch Set 2:

For me this change works.
I like checkboxes as much as Bradley :).

Also the idea of having them all checked by default for each transition that might happen is nice.

I've received notification when alarm transitioned to ALARM and did not receive one if I left ALARM notification disabled.

IMHO this is ready to be merged.

Patch-set: 2