Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

Besides Witold comments, I've tested this patch in my devstack environment and the 'deterministic' checkbox is present in the dashboard, but I just was not able to create an alarm, in both cases, deterministic being checked or not. Also, to make sure it was not a problem with my environment, I tested it without this change and I'm able to create alarms definitions normally.

Patch-set: 7
Label: Code-Review=-1