Update patch set 7
Patch Set 7:

I

 > > > Are you planning on introducing some validation to the form?
 > > > Currently it is easy to submit the form with empty values.
 > >
 > > Could you list those values, because by default there are few
 > > values that are actually required here (like name and
 > expression).
 > > Expression itself is at minimum composed of function, metric and
 > > threshold.
 > 
 > The values are in the expression part of the form. When you click
 > the plus next to the first expression for example three times, the
 > second expression is not initialized like the others.

I believe your comments are about compound alarms, not this change (multi step). Here forms acts pretty much as it was.

Originally if you open alarm definition form and underlying field expression is already populated with expression thus the only value that is not set and required is name.

Here it is similar, I think that we can think on expression validation in compound change and/or decide if we want to populate all inputs with default values.

FYI: This change is about transforming single step form into 3 step wizard and I'd like to stay on this path.

Patch-set: 7