Update patch set 9
Patch Set 9: Workflow-1

I think this change needs more work before it could be merged, right now it is possible to have invalid subexpressions in form and it is not picked during validation, setting that to -1 until resolved.

FYI: Change can be still tested and commented, changes to come will be about validation mechanism happening the background.

Patch-set: 9
Label: Workflow=-1