Update patch set 15
Patch Set 15:

(1 comment)

The visual difference between chevron and check icons is very subtle. I often tried to move the expression down and collapsed it instead. I think that the "down" button should be at the bottom.

Perhaps some tooltips would be a good idea too.

Maybe it would be a good idea not to collapse the above expression if the previously edited is not valid?

Patch-set: 15
Label: Verified=0