Update patch set 1
Patch Set 1:

Sounds good, Tomasz.

So, instead of creating a new checkbox, as I did, are you proposing to display an icon in the same box where we select the metric, right? And, maybe, a tooltip with 'regular metric' and 'sporadic metric' when the cursor is over the new icon. Is that what you were thinking?

Patch-set: 1