Update patch set 5
Patch Set 5:

Thanks for the review tomasz.  We started down this path because the update notification path is broken -- period must be provided on updated.  This patch works and fixes that issue, so I'd at least to get that working again.  If you want to enhanced this to by dynamic, that's OK with me.  I'll let others comment too.

Patch-set: 5