Update patch set 5
Patch Set 5:

> I think dynamic UI for notification is difficult. Because it maybe
 > needs angular code and I think there is no angular code for
 > notification. So to achieve dynamic UI we have to write many
 > modification.

Well, not that much actually, all you'd have to do is to ensure that all inputs will be finally rendered as if django would to that.

Django is great in providing UI from classes but that way it is sometimes very...how to put that...inflexible. Some of the forms, dialog windows from horizon core rely on custom HTML code wrapped with Angular from what I remember.

Patch-set: 5