Remove expression validation from UI
Expression validation available in UI
is not working well with compound
alarm expressions for action: Edit alarm definition.

Following commits removes that code and effectively
moves burden of such validation to the server side.

Additionally it changes UPDATE to PATCH as the method
to update alarm definition. With that only modified
fields are sent to the server

Closes-Bug: 1599027
Change-Id: If43ce287038724f3d39186c4b0209823a02757a6