Proxy cookies from and to Kibana
The fts-keystone plugin uses sessions to store
information about authenticated user. As session id
is stored in a cookie, this needs to be passed
between browser and Kibana, so monasca-ui needs
to proxy them. Without this, every action (request)
in Kibana would need to be auhenticated by Keystone.
This results in unnecessary load on Keystone
and also impacts performance.

Change-Id: Ie682443e87ee43fa30beee637b7bbf3873362f25