Update patch set 2
Patch Set 2:

Is this to supply the auth token from Horizon?

We discussed Kibana integration at the Summit, so you might want to check with Tomasz and Witek on this, but I'm OK with this change if they agree.

Note, we agreed to also support logging into Kibana similar to what has been done in Grafana, if possible. The main reason is that Horizon might not always be used and we want to allow users to use Kibana independently of Horizon.

Patch-set: 2
Label: Workflow=0