Add Constraints support
Adding constraints support to clients is slightly more complex than
services as the clients themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove monasca-ui.

Change-Id: Ie2f4e3ed3e309b1176ec8bffa5466da25345fd64