Update patch set 3
Patch Set 3:

Tomasz, the previous tox.ini file only enabled one check.  This is enabling many more checks and the other commits associated with this are fixing the scary checks we weren't running.  The remaining checks are mainly style based and not something we've really been concerned with on other repos.

Patch-set: 3