Update patch set 5
Patch Set 5:

> > Hi, I saw that you removed CHOICES.
 > > What is the way the translations are handled right now ?
 > 
 > You have sharp eyes.
 > Actually I have no idea to translate notification methods right
 > now.

Either we will just leave it for now and create bug @ launchpad (or storyboard if that is possible already) and try to fix that in future. I mean for english speakers that is hardly a problem, but I recall some work translating certain parts of monasca, so I think that someone will report a problem somewhere. 

In Openstack there are translations done in :

- client side of horizon
- python part of horizon
- exception messages can also be translated

so maybe if we would take oslo.il8n library and do the translating in the monasca-api endpoint that would be fine enough (the only requirement would be to add some locale files there with translated strings), but again that would be just for python part.

I cannot say if I am fine with not having this or not, maybe that's point worth discussing during the weekly meeting.

Patch-set: 5
Label: Verified=0