Fix monasca-ui forms
Fix problem with unexpected keyword argument 'name' for
build_attrs() from Django. 'name' was passed to the method
as a kwargs. **kwargs in build_attrs() were removed.
Pass 'name' in a dict.
In Django 1.10+ passing Context to render is deprecated and
gives an error, pass instead a dict.

Change-Id: I2d39a10f08ac60c163f80cca3256a2bb24b4e0a2
Story: 2001133
Task: 4844