Allow raw Grafana links
This commit introduces the GRAFANA_LINK_STYLE setting which
allows the user to point the dashboard URLs specified through
GRAFANA_LINKS at arbitrary paths or URLs without further
modification. For this to happen, GRAFANA_LINK_STYLE needs to
be set to 'raw'. The default setting of 'names' will cause the
links to be assembled as they were assembled before this
commit.

Change-Id: I55660f91fd56eb22dd7ad4674af422aeb2ba2ea3
Story: 2001305
Task: 5870