Update patch set 1
Patch Set 1:

There is a problem with validation of empty CharField in Django version used in stable/ocata. White spaces are treated as a regular string. For a later version, it is fixed and strim() is enough. Here I had to add additional validation for white spaces.

Patch-set: 1
Label: Verified=0