Update patch set 1
Patch Set 1:

Thanks @Withold, Our QA found an issue in Horizon of our cloud which is pike based. He created a regular user, this user can see the monitoring panel but can not see the data. Whenever try to access data it throws exception. The monitoring panel should not be visible to a regular user. I applied the master's policy to the pike based cloud, it seems to work.  I didn't realize some compatibility issues in monasca-api. Please advise how to proceed to fix pike and queens problem. Thanks.

Patch-set: 1